+++ src/text.hpp
@@ -15,13 +15,13 @@ public:
 
 	void update();
 	void setContent(const string & value);
-	void setOpacity(const float value);
+	void setOpacity(const double value);
 	void setColor(const fvec3 & value);
 	void setPosition(const fvec2 & value);
-	void setRotation(const float value);
+	void setRotation(const double value);
 	void setSize(const fvec2 & value);
 	void setVisible(const bool value);
-	void setLightness(const float value);
+	void setLightness(const double value);
 
 	const vector<shared_ptr<Quad>> & getQuads() const;
 
@@ -32,9 +32,9 @@ public:
 	const fvec2 & getPosition() const;
 	const fvec2 & getSize() const;
 
-	const float getOpacity() const;
-	const float getRotation() const;
-	const float getLightness() const;
+	const double getOpacity() const;
+	const double getRotation() const;
+	const double getLightness() const;
 
 	const int getDepth() const;
 
@@ -157,9 +157,9 @@ private:
 	fvec2 _position = fvec2(0.0);
 	fvec2 _size = fvec2(1.0);
 
-	float _rotation = 0.0;
-	float _opacity = 1.0;
-	float _lightness = 1.0;
+	double _rotation = 0.0;
+	double _opacity = 1.0;
+	double _lightness = 1.0;
 
 	bool _isVisible = true;
 };
