+++ src/timer.cpp
@@ -34,7 +34,7 @@ void Timer::stop()
 	_clock->isStarted = false;
 }
 
-const float Timer::getDeltaTime() const
+const double Timer::getDeltaTime() const
 {
 	if(_clock->isStarted)
 	{
@@ -42,7 +42,7 @@ const float Timer::getDeltaTime() const
 	}
 
 	const long long difference = (_clock->stopEpoch - _clock->startEpoch);
-	const float deltaTime = static_cast<double>(difference) / 1000000.0;
+	const double deltaTime = static_cast<double>(difference) / 1000000.0;
 
 	return deltaTime;
 }
